[BUG] Prevent unconditional unwrap of collection config (#4404)
## Description of changes

Previously, the frontend was unconditionally attempting to unwrap the
collection config, even if it was None. This wasn't caught before
because all of the popular clients are actually passing in a collection
config. However, it should be optional.

## Test plan

Not tested yet, will check in staging after it merges.

## Documentation Changes

n/a