[CLN] Fix Port Parameter Inconsistency in HttpClient (String to Integer) (#3662)
## Description of changes

*Summarize the changes made by this PR.*
- Refactored Port Parameter Handling: Changed the port parameter in the
[HttpClient
class](https://github.com/chroma-core/chroma/blob/479ce3312c18545f6d5d741b28f0b45b97ea391a/chromadb/__init__.py#L164)
to accept an integer instead of a string. This applies to both the
Python code and the related documentation.
  - Before: port="8000" (string)
  - After: port=8000 (integer)
- Docstring Updates: Updated the docstrings in __init__.py to reflect
the new default port value as an integer (8000 instead of "8000").
- Documentation Fixes: Corrected the documentation in client.md to
reflect the updated data type for the port parameter, ensuring
consistency with the code changes.

*Motivation*
I was interested in your project and started reading through the source
code with the intention of running it and potentially contributing.
While setting up the environment, I noticed an inconsistency: the
default port was defined as a string ("8000") in some places, while the
interface expected it as an integer. I thought this would be a good
starting point for my contributions, addressing a minor but important
issue to improve code consistency and reliability.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

- Docstrings Updated: Yes, all relevant docstrings in __init__.py have
been updated to reflect the changes to the port parameter.
- External Documentation: Yes, updates were made to the client.md file
to align with the code changes.
- Docs Repository Update Needed: No additional changes are required in
the [docs repository](https://github.com/chroma-core/docs) as the
in-repo documentation has been updated accordingly.