[BUG] Handle duplicates in chroma-load.
I'm seeing the error, "invalid request: IDs lengths are different: got 9, expected 10".

I assume this is from the verifying data set as it's the only place this
string could be sprintf'd.

This de-dupes keys before fetch to avoid the service doing the deduping
and returning a short set.  I wasn't able to repro locally, but am in
staging, so this will pass tests and then get debugged in staging.