[ENH] CLI Browse command (#4431)
## Description of changes

Adding the `chroma browse [collection name]` command for a TUI app to
browse collections.

Arguments:
* `collection_name` (required) - the collection to browse
* `--local` - find the collection in a local Chroma server. The CLI will
try to connect to a Chroma server on localhost:8000, or stand up a local
Chroma frontend if `chroma.config.yaml` is found in the current working
directory.
* `path` - data path for local Chroma.
* `config` - path for a config file for local Chroma.
* `host` - Chroma host with the collection.
* `db` - the name of the DB with the collection.

If "local" arguments are not provided, the CLI will assume this is a
Cloud collection. If in this case `db` is not provided, the CLI prompts
the user for selecting the DB with the collection.

The code for the TUI app is under `rust/cli/tui/collection_browser`. The
main files are:
* `app_state.rs` - defines the state of the app. It also has functions
that handle actions sent by the event handler, that should modify the
state of the app.
* `events.rs` - contains the event handling code in response to
different key events, using a sender-receiver pattern. The `next` method
uses `tokio::select` to handle both UI events and asynchronous
operations (loading records, submitting queries). This pattern allows
asynchronous operations to not block the UI.
* `mod.rs` - contains the main app loop. On every iteration, we handle
incoming events, fire asynchronous tasks, and draw the UI.

## Test plan
- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
In a separate PR