[CLN] Warn on log pull failure instead of error (#4435)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- There are valid cases for this log pull to fail, it is not an
invariant violation as stated.
- If the log is failed to be updated with the latest compaction offset
(Which is best-effort), but the sysdb was, the log may schedule a
collection for compaction that has no records outstanding. This leads to
hitting this case, which is valid. In the future we could repair the
log.
- Forked collections will also run into this as the fork() call may copy
the old log offset, but a compaction happens after, this leads to them
disagreeing but progression is always forward, it manifests as the same
issue above which is that the log is behind the sysdb.
- Keeping this as a `warn` because its worth monitoring this as it
implies log service may be failing to get updated. Or that the fork race
is commonly copying unnecessary data.
- New functionality
  - None

## Test plan
_How are these changes tested?_
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None