[TST] Debug cli tests (#4453)
## Description of changes

Speed up python CLI wrapper tests:
* Run tests directly instead of setting up the integration test
environment.
* Previously, vacuum set its own timeout for SQLite connections, however
now we have a default timeout of 1000 seconds (~17 minutes) on all
connections. The last test in `test_cli` checks that we abort long
standing vacuum operations (when the DB file is locked). To work around
the 1000 timeout, added a `timeout` argument to the CLI command to
override the default long timeout. The DB set up by the test would have
been vacuumed in a matter of seconds.

## Test plan
- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
N/A