[ENH] add example of forking using github repos (#4413)
## Description of changes

This PR adds an example notebook to show forking usage with github repos
and branching. It creates a custom chunker using tree sitter to generate
embeddable chunks of code, and a github helper to pull file info, track
diffs between branches. it then uses chroma cloud to index and adds a
github repo into a collection, and based on the branch provided creates
a fork of the collection, calculates the diff between the original
branch and the new branch, deletes chunks from the new branch that
correspond to the diffed files, and rechunks and reinserts them.

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_