[BUG]  Catch a bug that causes wal3 to hang under contention. (#4476)
## Description of changes

If two wal3 instances try writing the same log, only one will succeed,
and the other will fail.  This can, under rare situtation, lead to a log
failing recovery due to contention, and thus failing to open.
Contention should be a transient thing and the log should handle it.

Included is a claude-written test that plays ping-pong between two log
writers in a coordinated fashion.

## Test plan

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A