[ENH]  Add caching to rust log service. (#4465)
## Description of changes

Using foyer, add caching to the rust log service.  This caches parquet
fragments listed to be fetched to complete a call to pull_logs.

Caching isn't used for the dirty log or forking logs, just pull logs.
It gets populated by a key that is (collection_id, path), and then gets
immutably served from that point forward.

## Test plan

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A