[BUG] Retry chroma-load upserts when rate limited
When the upsert operations are rate-limited, the "block" of 100 records in the
operation is never inserted. Therefore, the cardinality of the test data set
can never grow beyond the point of the failed (rate limited) upsert.

Also, when starting a new workload, we need to reset the cardinality and heap
for the data set, so that it starts again from cardinality of 0.