[BUG] Retry chroma-load upserts when rate limited (#4485)
## Description of changes

When the upsert operations are rate-limited, the "block" of 100 records
in the operation is never inserted. Therefore, the cardinality of the
test data set can never grow beyond the point of the failed (rate
limited) upsert.

Also, when starting a new workload, we need to reset the cardinality and
heap for the data set, so that it starts again from cardinality of 0.

## Test plan

Tested locally in tilt. Verified that the cardinality of the test data
set continues to grow, even after hitting rate limit errors on previous
upsert attempts. Also verified that the cardinality is reset to 0 when
initializing a new workload.