[ENH] write embedding function to config if provided (#4496)
## Description of changes

We had previously decided that providing embedding functions would now
be done via the config. This PR makes it so that embedding functions,
even if provided via the existing embedding function parameter, would
still be routed to the config in the case where the user does not
specify an EF in the config. So priority would be
1. config provided EF
2. parameter level EF

- Improvements & Bug fixes
  - ...
- New functionality
  - ...

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_