[BLD]: unify Rust Dockerfiles for faster builds (#4479)
## Description of changes

This PR unifies the 5 separate Rust Dockerfiles into 1. This results in
substantially faster builds as well as being cleaner and easier to
maintain.

After this PR, the old Rust Dockerfiles are no longer used in this repo
but I'm keeping them around until all downstream consumers are updated
to use the new `rust/Dockerfile`.

## Timings

I tested the difference in (local) speed with 3 scenarios. For each one,
I measured the time from the start of the build to when all services in
Tilt were green.

| | before | after | speedup |

|----------------------------------------------------|--------|-------|---------|
| cold `tilt up` | 7m14s | 3m55s | 1.8x |
| new comment in `rust/types/src/api_types.rs` | 2m2s | 58s | 2.1x |
| new comment in `rust/garbage_collector/src/lib.rs` | 37s | 29s | 1.3x
|