[ENH]  Support custom data sets for chroma load. (#4547)
## Description of changes

This exposes the already-existing ability to shove a data set into
chroma-load.  The `/start` endpoint now takes a json blob xor a data set
name.  If the `/start` endpoint is given just the name, it resorts to
the old behavior so as not to break anything.

If given a JSON blob, it will interpret that as the data set and try to
run it.  Right now the only supported workloads would be to pass a blob
that's pre-configured, or to pass a `References` struct serialized to a
data set.  The chroma-load-start command has been updated to facilitate
this.

## Test plan

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

- [X] --help for chroma-load-start has been updated