[BUG] Restore increase max payload size of log service (Go) (#4546)
## Description of changes

This restores the core functionality we were hoping to add in #4534
which is increasing the gRPC payload size. Instead of using the methods
from the `grpcutils` package, this just sets the option in the existing
setup so that we hopefully don't break whatever voodoo is keeping
metrics collection working.

- Improvements & Bug fixes
  - See #4534
- New functionality
  - See #4534

## Test plan

I ran `make test` locally, the build succeeds, but our local tests are
broken for some reason. Relying on CI so that I don't have to debug yet
another orthogonal issue.