[ENH] use subqueries for fts locally (#4536)
## Description of changes
previously on FTS operations, the left join + filter meant all rows were
accumulated before applying the where clause to limit the results. Now,
we introduced a subquery approach, similar to metadata so that the
subquery can resolve first, and removed the left join. Since we are now
using in on the subquery, and embedding<>metadata are 1:1 (unlike
embedding<>metadata which is 1:many meaning we must do the left join).
the left join on the fts table gets optimized away by sqlite anyways, so
this is a code clarity update to remove the lj.
before:

![image](https://github.com/user-attachments/assets/8862776f-2b60-4544-b31c-981f77070c69)
after:

![image](https://github.com/user-attachments/assets/7c8ff5c9-5424-4c68-ae85-d2387c6a27c1)


## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_