[BLD] Do not trigger deploy on merge to release branches (#4561)
## Description of changes

This removes the `release/**` trigger from the `trigger-deploy` workflow
so that deploys aren't created when hotfixes are merged to release
branches. We want this to be a manual step for the time being to prevent
unintentional deploys of the control plane when we are hotfixing the
data plane.

- Improvements & Bug fixes
- Prevents hotfix merges from triggering a deploy of code that is
potentially behind what is deployed.
- New functionality
  - N/A - removes automation

## Test plan

Manual testing -- we will have to manually deploy changes after merging
hotfixes to release branches.