[ENH]  Add a safety cutoff to the rust log service. (#4535)
## Description of changes

If a collection on the rust-based log has more than a
configuration-configurable number of records, shut it down hard no
matter what.  This happens when compaction gets backlogged.  By default,
no more than 1_000_000 can be on the log uncompacted.  This is just a
placeholder that we can change.

## Test plan

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A