[CHORE] Remove validation on ef config update for model path (#4571)
## Description of changes

Some embedding functions allow the user to pass a path for their
embedding function. We should not be validating that the user is trying
to update the model path, as that could be too restrictive.

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_