[ENH] Bootstrap a wal3 log from existing content. (#4560)
## Description of changes

This adds a "bootstrap" call to wal3 that copies the data
semi-atomically. It is reasoned to be
safe to do concurrent with all other log operations, but may leave a
FragmentSeqNo(1) lying around
in the event that a log is initialized at the same time it's
bootstrapping. This should never
happen in our system so thankfully we can just leave the hole.

See #4558 for the reasoning.

## Test plan

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A