[BUG] Fix race in get_or_create (#4568)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- Fixes a race in get_or_create where a deleted collection may be
returned, or where get_or_create may be inconsistent between collection
and metadata, or where you may return a unique constraint violation
despite get_or_create semantics. The details of the fix are well
commented in the code and I delegate detail to those comments.
- I added `ChromaError` to the chroma types, for generic error handling
fallbacks
- This error is propagated all the way through and I validated all cases
will be properly reported when trigerred.
- New functionality
  - None

## Test plan

_How are these changes tested?_
I added a test that stresses the race, it fails on main.
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None