[ENH] Change up the alt-host shunt for log service. (#4588)
## Description of changes

The log config now takes a UUID.  Every UUID <= to that UUID (so that
ff... will be repesentable) will be forwarded to the rust log prior to
the go log.  Every UUID > the configured UUID hits the legacy service.

This is so that we can incrementally move production traffic to the rust
log service from the go log service.

## Test plan

I wrote a test of the logic for which UUIDs are on the log.

The rest is in CI's hands.

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A