[ENH]  Add a way to seal a tenant's Go log immediately upon collection creation. (#4603)
## Description of changes

I want a way to make a tenant bound to the new log for all collections,
no matter when they are created.  One way to handle this would be to
make the log service tenant-aware and have it do the selection
intelligently on the fly.  This would require threading the tenant
information in the proto and trusting it all the way down.

The alternative taken in this PR is to deal with it entirely from the
frontend by sealing the collection on the critical path of collection
creation, opting to trace an error rather than fail the op entirely if
sealing the log fails.

## Test plan

CI

## Documentation Changes

N/A