[BUG]  Sealing a collection right away fails. (#4602)
## Description of changes

Because there would be no row in the DB, sealing the collection fails to
update is_sealed to true.   Add a fallback to insert a sealed row if the
update fails.  Add a test to capture that this works.

## Test plan

CI.  Ran the new test locally.

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A