[BLD]: fix: set default channel of rust from rust-toolchain.toml (#4612)
## Description of changes

The `build-sdist` release job [is
failing](https://github.com/chroma-core/chroma/actions/runs/15193462410/job/42739524860)
because it's using the version of Rust provided in the runner image
(1.80) instead of the version we install with rustup (1.81). This change
sets the global default to 1.81, from our `rust-toolchain.toml`, so it
should resolve the build issue.