[PERF] Tune keepalive for chroma cloud (#4618)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- By default Chroma Cloud allows keepalive (we use HAProxy and it is on
by default) and uses the http-request timeout if the keepalive timeout
is not set - which for us is 50s. This raises the keepalive timeout from
default 5s to < the haproxy timeout. We choose to not make this
configurable for now, since the bottleneck will be the downstream
config. This at least allows reusing the connection in a larger window
- New functionality
  - None

## Test plan
_How are these changes tested?_
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None