[BUG]: root collection ID should be null if empty string (#4663)
## Description of changes

Some rows in our database have `root_collection_id: ""` instead of
having it set to null. We're not sure why this happened, this is a quick
patch as a short-term fix while we continue to investigate.

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_