[BUG] Raise Error when can't deserialize configuration json from server, lazily load ef on CollectionModel, warn on api_key (#4471)
## Description of changes

This PR fixes a bug where it warns users if the configuration is not
deserializable rather than raising an error, giving the users an empty
config. Now, it raises an error if the configuration is not
deserializable. This will only happen when the client is ahead of the
server, which we do not need to support.

EF's from the config are now lazily loaded. only when ._embed is invoked
from the model is it tried to be built, and will take precedent over the
provided ef.

This also fixes a bug with list_collections, where if the user tries to
load collections without the proper embedding function parameters set
up, like api key env var, etc it would throw an hnsw and spann not
configured error, which is not the root cause. instead, it will now load
only the configuration_json, and at embed time try to load the ef, so
errors will only propogate at that point, and with much clearer error
messages.

Now, using api_key directly in embedding_functions will give a
deprecation warning saying it is not persisted.

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_