[PERF]: skip document mutation in full-text index writer if old and new documents are same (#4678)
## Description of changes

If the old and new documents are identical, we can save compute by
ignoring the update.

A customer on Chroma Cloud seems to have a workload with a lot of no-op
updates, so this should improve compaction speed for their collections.

## Test plan

_How are these changes tested?_

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_

n/a