[ENH]  Do not keep around logs that get passed through.
There's an explicit assumption in wal3 that logs will be bound to a
single writer.  To make concurrent writes possible we use the state hash
table to get to a single log.

If there was a case of serial writes, we would end up creating an entry in
the hash table, do the write, and then tear down the entry in the hash
table as part of the RAII-driven return.  This means every operation
would be a re-opening.

Instead we keep the logs around for 60 seconds.  This could be made
configurable.

This PR makes it so we only keep logs around for collections that are
already initialized.