[ENH]New JS Client (#4632)
## Description of changes

New implementation for the Chroma JS/TS Client.

### Project Structure

The project is structured as a monorepo using `pnpm` workspaces:
* `packages/chromadb` - contains the client code.
* `packages/ai-embeddings` - contains packages for different embedding
functions, one for each provider.

### Embedding Functions

The Chroma JS/TS client now includes a modular system for AI embedding
providers, located in `packages/ai-embeddings`. Each provider is its own
package (e.g., `@chroma-core/openai`), and the `@chroma-core/all`
package which re-exports all others for users who want access to all
providers at once. All embedding functions have been updated to use the
latest clients and up-to-date models and API arguments for each
provider.

Each package contains its own build and test suites, and lists
`chromadb` as a peer dependency. When an embedding package is installed,
it will call the `register` function from
`chromadb/src/embedding-function.ts` to register itself as available for
the client. For example, to get access to `OpenAIEmbeddingFunction`,
users need to install `@chroma-core/openai`.

The `@chroma-core/default-embed` is the new implementation for our
default embedding function. Instead of using a fork of
`huggingface/transformers` we now use it directly as a dependency, which
makes it easier to work around errors we have seen in the past with
`chromadb-default-embed`. Whenever a user creates a collection without
providing an EF, `chromadb` will try to dynamically import
`@chroma-core/default-embed` and error if it is not installed. Upon
successful import, we register the default EF. This logic is
encapsulated in `getDefaultEFConfig`.

Another utility function for embedding functions is
`getEmbeddingFunction`. Given an EF configuration (from a collection
configuration), it returns an embedding function object. If the EF
config is of type `legacy`, we will try to return the default EF.
Otherwise, the EF config specifies an EF from a provider or a custom
one. If the name of the EF from the config is not registered with the
client, we error. For example, if the EF config specifies `openai`, but
`@chroma-core/openai` is not installed (and hence not registered), we
can't construct the `OpenAIEF`. If users use custom EFs, they need to
register them manually. If the construction of a registered EF failed,
we error. Note that on all error cases from `getEmbeddingFunction`, we
return a `MalformedEF`. This is a placeholder object that allows the
collection to be used for operations that don't require `.embed(...)`.
The `MalformedEF` saves an error message with why the construction
failed, and will throw it if users use `.add` or `.query`.

### API Generation

Previously the API generation required a running Chroma server locally,
and involved a series of scripts. It also produced incomplete types. We
now use the Rust binary to start a local Chroma server (as the tests
do), and an updated script in `chromadb/scripts/gen-api.ts`.

All the `_Client` classes hold an instance of the generated API client,
and pass to it a custom `fetch` implementation (in `chroma-fetch.ts`),
which maps error codes to custom client errors.

### Client-Side Validations

The client nows performs client-side validations that are identical to
those of our Python client. This should make it easier to keep this part
of the code in-sync with the Python code.

### `ChromaClient`

Previously, `ChromaClient` called `getUserIdentity` on every function
call to resolve and set `tenant` and `database` from an API key in case
they were not provided or verify the provided ones.

You will find a call to `getUserIdentity` in the classes's `_path`
method. It is used by all functions to provide the generated API client
the correct endpoint path (i.e. `tenant/database/...`). However, it will
only call `getUserIdentity` to resolve the `tenant` and `database` name
, if they were not provided, and `ChromaClient` will always set them by
default to `default_tenant` and `default_database` respectively. The
`CloudClient` (see bellow) will override these default setting if
`tenant` and `database` were not provided.

Additionally, the client construction now mirrors that of the Python
client:

```javascript
const client = new ChromaClient({
      ssl: true,
      host: 'api.trychroma.com',
      tenant: 'tenant',
      database: 'database',
      headers: {
        'x-chroma-token': 'API KEY'
      }
});
```

### `CloudClient`

We are adding a `CloudClient` (and `AdminCloudClient) to JS/TS users. In
addition to `tenant` and `database` it takes an `apiKey`. If only an
`apiKey` is provided, we automatically resolve `tenant` and `database`
using `getUsingIdentity`. The `CloudClient` can be instantiated with no
or partial arguments if the user has set the `CHROMA_TENANT`,
`CHROMA_DATABASE`, and/or `CHROMA_API_KEY` environment variables.

The `CloudClient` extends `ChromaClient` and sets `host`, `ssl`, `port`,
and `headers` correctly to connect to Chroma Cloud.

### Collection Types

The `CollectionAPI` interface includes `id`, optional
`embeddingFunction` and all the Collection API functions. The
`Collection` interface extends `CollectionAPI` and defines `name`,
`metadata`, and `configuration`. We only export these interfaces from
`collection.ts`. In `collection.ts` we have the `CollectionAPIImpl` and
`CollectionImpl` implementing these interfaces.

The `CollectionAPI` interface is introduced to allow users to get access
to collection API functions without having to get a collection with
`.getCollection`.

`.getCollection` works as before, returning a `Collection`
(`CollectionImpl`). We introduce the `.collection()` function, which
takes in a collection ID and an optional `embeddingFunction`, and
returns a `CollectionAPI`. This client function is introduced to support
users in serverless environments, where `Collection` objects don't
persist, so they can discount the API call of `.getCollection`.

### Collection Getters

We add the `getCollections` and `collections` client functions. These
take in a list of collection names or IDs respectively to allow users to
get several collections at once. These can then be de-structured:

```javascript
const [col1, col2] = client.getCollections(["col1", "col2"]);
```

### Row Results

`GetResult` and `QueryResult` are now classes that we construct on
`.get` and `.query`. Their fields are the columnar representation of the
results of these operations, and they now also include a `rows()`
method, that returns the row-based version of the query. For example, a
`GetResult` object has fields `ids`, `documents`, etc. And the `rows()`
method on that object returns the following:

```javascript
{
   include: Include[];
   records: { id: string; document: string; ... }[]
}
```

### Metadata Types

Users can now provide a type to `.get` and `.query` to define the type
of the returned metadata from these operations. This is meant to be a DX
improvement to allow autocompletion and avoid type checking for
metadata:

```javascript
const { records } = collection.get<{ metadataField1: string; metadataField2: number }>(...).rows();
console.log(records[0]?.metadata.metadataField1)
```

### NextJS Support

We have seen that webpack fails in NextJS with the default embedding
function due to its dependencies on some bindings. Previously we
instructed users to add some configurations to their `next.config` file
to resolve these issues. The new client contains a `withChroma` plugin
that users can import into their `next.config`, which includes these
configurations.

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

In a separate PR