[ENH]  Make wal3::boostrap robust to zero records and fix offset conditions in log (#4702)
## Description of changes

The implementation of wal3::LogWriter::bootstrap was broken when there
were no log records.
Introduce an initial_offset to the manifest so that we can bootstrap
empty logs.

Also, change up the ordering. There was a race where a log could be
initialized to zero, read by
someone else, and then bootstrapped; only put the correct manifest after
initializing any fragments.

This also adjusts offsets in the log to make tests pass for a PR up the stack.

## Test plan

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A