[ENH] Limit the number of records pulled by pull logs (#4710)
## Description of changes

Here's the scenario:  A client is writing in batches to the log.  The
batches come in groups, and wal3 does de-amplification of batches.
Imagine running a pull logs.  Prior to this change, the pull logs can
only knows which files to fetch based upon the number of files or number
of bytes.

By pruning to number of records, we allow for the case where there's one
fragment that has BatchSize records, and so we only need to fetch one
fragment to read its records.

## Test plan

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A