[BUG]  Seal collections on fork in the go log. (#4767)
## Description of changes

If the collection is sealed, and it gets forked, the new collection
should be sealed too; otherwise we leak wal3 tenants back to the
go-based log.

## Test plan

I'm sending it to ci.  test_fork.py fails locally, but it seems to be
failing by trying to fork a collection that doesn't exist, which makes
no sense.  Sharing to debug.

## Documentation Changes

N/A