[TST]: fix flakes in Rust tests caused by port conflicts (#4788)
## Description of changes

https://github.com/chroma-core/chroma/pull/4780 started running the Rust
tests in parallel in CI, which resulted in a new type of flake. The
`random-port` crate is not safe to use across processes without
additional locking/retry logic as it does not actually bind to and lock
the port. Because of this, tests occasionally flake with port conflict
errors when run in parallel.

This change fixes the issue by acquiring a random port at bind time.

## Test plan

_How are these changes tested?_

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust