[TST] Move windows-based tests out of PR Checks (#4748)
## Description of changes

Move our Python+Windows tests to a separate workflow that only runs on
merges to main.

- Improvements & Bug fixes
- This should make our PR checks faster by removing the slowest part of
them
- This fixes a few potential flakes in the Windows tests that were
discovered when running a large number of Windows tests concurrently
- New functionality
- The `_python-tests.yml` workflow is now parameterized over a `runner`
input to specify the runner for the jobs
- The `pr.yml` / "PR Checks" workflow runs _most_ tests against against
Linux/Ubuntu and adds a new "python-smoke-tests" job that runs the
`test_api.py` tests against Windows when the `runner` is not Windows

## Test plan

The PR checks should pass.

---------

Co-authored-by: Max Isom <codetheweb@users.noreply.github.com>