[ENH]  Tool to purge a collection from the dirty log. (#4787)
## Description of changes

This affects wal3/rust-log-service.  When a collection is on the dirty
log it can be scheduled for compaction.  This tool forcibly inserts a
"purge" marker to the dirty log that effectively drops the marked
collection from the dirty log until it gets written to again.

Use with care.

## Test plan

I'm relying upon the compiler and visual inspection of the code.

## Documentation Changes

N/A