[ENH]  Add Copy API to chroma storage. (#4798)
## Description of changes

This API will use the serverside option to upload a file, meaning that
copying a file becomes a cheap operation, relatively speaking.  No
longer will it have to stream to the rust-log-service and back to S3.

## Test plan

I added a unit test.  CI for the rest.

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A