[BLD]: remove architecture specification from actions/setup-python (#4816)
## Description of changes

See https://github.com/chroma-core/chroma/pull/4815 for context on the
issue. That PR did not fix the issue because the error was actually
caused by an incorrect architecture argument, e.g. `x86_64` was
specified but the action expected `x64`. The action defaults to the
architecture of the runner, so we don't need to manually specify it.