[ENH]  Support list_prefix operations for S3 and AC/S3. (#4637)
## Description of changes

This adds support for a list_prefix operation that calls the V2
ListOBjects API in the AWS SDK.  I've included an integration test.

The intended consumer of this API is the wal3 garbage collector,
which needs to list cursors in the directory to find the lowest-version
cursor.

## Test plan

Integration test added.

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A