[BLD] build arm64 image on arm runner instead of in QEMU (#4792)
This PR re-adds split builders for image releases so that ARM images can
be built on native platforms. The issue before was that we ran
`useblacksmith/build-push-action@v1.1` twice in the same workflow, which
led to undefined behavior.

Validated that the new image build workflow succeeds by running in this
PR (although I can't test the manifest merging until a run on main).