[ENH]  Switch garbage collector off object-store-based backend (#4846)
## Description of changes

The best intentions of mice and men or something led us to have an
ObjectStore backend, and the path
of least resistance led to us using it in tests.

Change all instances of that backend to use S3 storage with a Minio
credential.

## Test plan

CI
    
- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A

---------

Co-authored-by: macronova <trivial@invariantspace.com>