[ENH]  Reduce log contention when writing concurrently. (#4842)
## Description of changes

The LogContention Error could manifest if two concurrent writers or a
writer and a garbage collector race.  This would fail CI cluster tests.

To remedy this problem, this PR reworks the retry logic to make sure
that when there is contention the writers will always reset to a good
place.  It does this by making a pair of writers alternate.  If either
writer gets stuck the test fails.

## Test plan

CI

## Documentation Changes

N/A