[CHORE] Add base64 encoding on preflight checks (#4843)
## Description of changes

This PR removes the base64 setting, and replaces it with a preflight
check that the server has base64 enabled, and if it does will use it.
The preflight check is cached on the client for reuse.

It also adds preflight checks to the js client, caching it on the client
for reuse.

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_