[ENH]  A scrubbing tool that supports limits.
It'll scrub the first so many bytes and files of the manifest.  In the
event that the entire log can be read within the limits, it'll compare
the setsum of the read files to the setsum of the manifest and return
that.  In the event of a short read, it cannot compare the manifest, but
will return an error indicating a short read and will not return a
setsum.