[ENH]  A scrubbing tool that supports limits. (#4891)
## Description of changes

It'll scrub the first so many bytes and files of the manifest.  In the
event that the entire log can be read within the limits, it'll compare
the setsum of the read files to the setsum of the manifest and return
that.  In the event of a short read, it cannot compare the manifest, but
will return an error indicating a short read and will not return a
setsum.

## Test plan

Tests pass locally. Will run in CI. Tool tested against tilt w/
chroma-load.

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_