[ENH]  Add a path to bail from snapshot, manifest, and gc installs.
There was a loop that would loop until success.  This could cause the
system to spin when Amazon has a problem.  All the code that calls this
is robust, so it should return an error after three attempts and let the
higher level code retry.  99 contention tests for two writers that
contend.