[ENH]  Add a path to bail from snapshot, manifest, and gc installs. (#4893)
## Description of changes

There was a loop that would loop until success.  This could cause the
system to spin when Amazon has a problem.  All the code that calls this
is robust, so it should return an error after three attempts and let the
higher level code retry.  99 contention tests for two writers that
contend.

## Test plan

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A