[ENH]  Allow garbage collection to provide a phantom cursor to retain more of the log.
Add a new argument to the garbage collect path that acts like an
additional cursor was temporarily placed on the log for the interval of
garbage collection.  The practical upshot of this is that garbage
collection can keep arbitrary amounts of data, but wal3 will guard
against garbage collection ever collecting past any cursor.

There's a TODO in the garbage collector to thread the appropriate offset
to the garbage collector.