[ENH][wal3]  Make GC be delete-free for log/gc/GARBAGE
AWS doesn't support delete of things with an If-Match header except in
directory buckets and we are not using those.  Instead overwrite using
the same hand-over-hand pattern as the manifest.  Load a GC, decide
what to do while holding on the the ETag, and then overwrite the garbage
with a sentinel rather than deleting it.