[ENH][wal3]  Make GC be delete-free for log/gc/GARBAGE (#4904)
## Description of changes

AWS doesn't support delete of things with an If-Match header except in
directory buckets and we are not using those.  Instead overwrite using
the same hand-over-hand pattern as the manifest.  Load a GC, decide
what to do while holding on the the ETag, and then overwrite the garbage
with a sentinel rather than deleting it.

## Test plan

CI

- [wal3] Tests pass locally with `pytest` for python, `yarn test` for
js, `cargo test` for rust

## Documentation Changes

N/A