[BUG]  The condition for setting the manifest initial_seq_no was broken.
The condition within wal3 for setting the initial_seq_no was broken.

The simpler condition is to set it to limit whenever limit is non-zero.
The start will always be <= limit by the guard at the top of the
function.  Therefore the only case this doesn't set the limit is when
limit is FragmentSeqNo(0), in which case we want an initial offset of 1.